refactor(views): mover views do chamado para app chamados
